dossier.web grows another tentacle.
The caller can now pass a pair of functions that know how to convert
to/from DB ids and user visible ids. These functions are used in almost
every web service. Each service must be careful to convert incoming ids
to DB keys and convert outcoming ids to user visible keys.

This is *not* a breaking change because these functions default to the
identity function.